Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -6,7 +6,7 @@ int main(){
     scanf("%d", &a);
     scanf("%d", &b);
     scanf("%d", &c);
-    float result = (a+b+c)/3.0;
+    int result = (a+b+c)/3.0;
     printf("Average: %.2f", result);
     return 0;
 }
